+++ lib/src/_src.g.dart
@@ -10,5 +10,9 @@
 // ignore_for_file: directives_ordering
 
 export './ansi_styled_string.dart';
+export './boot_audit.dart';
+export './boot_audit_storage.dart';
+export './in_memory_boot_audit_storage.dart';
+export './local_storage_boot_audit_storage.dart';
 export './log.dart';
 export './log_item.dart';